move all new header components to live in the same place as other existing header components
