+++ apps/src/code-studio/components/header/LevelBuilderSaveButton.jsx
@@ -7,7 +7,7 @@ import {
   setProjectUpdatedError,
   setProjectUpdatedSaving,
   setProjectUpdatedSaved
-} from '../headerRedux';
+} from '../../headerRedux';
 
 // Levelbuilder-only UI for saving changes to a level.
 class LevelBuilderSaveButton extends React.Component {
