+++ apps/src/code-studio/components/header/MinimalProjectHeader.jsx
